109550083 Lab4

+++ kernel/src/os/shell.rs
@@ -1,23 +1,49 @@
 use super::file_system::cpio;
 use super::stdio::{get_line, print};
+use crate::os::stdio::{println, println_now};
+use crate::os::timer;
 use crate::println;
+use core::alloc::Layout;
 use core::arch::asm;
+use core::time;
+use alloc::alloc::alloc;
+use alloc::string::{String, ToString};
+use alloc::boxed::Box;
+mod commands;
+use super::allocator;
+
+fn print_time(time: u64) {
+    let sec = time / 1000;
+    let ms = time % 1000;
+    println!("Time: {}.{:03} s", sec, ms);
+    // set_next_timer(2000);
+}
+
+fn set_next_timer() {
+    // println!("Timer expired");
+    timer::add_timer_ms(2000, Box::new(|| set_next_timer()));
+    // for c in b"Timer expired -- Function\n" {
+    //     unsafe {
+    //         crate::cpu::uart::send(*c);
+    //     }
+    // }
+    // loop {}
+}
 
 pub fn start(initrd_start: u32) {
     let mut inp_buf = [0u8; 256];
 
     loop {
         print("> ".into());
-        let _len = get_line(&mut inp_buf, 256);
+        let len = get_line(&mut inp_buf, 256);
         let initramfs = cpio::CpioArchive::load(initrd_start as *const u8);
 
+        if !inp_buf[0].is_ascii_alphabetic() {
+            continue;
+        }
+
         if inp_buf.starts_with(b"help") {
-            println!("help\t:print this help menu");
-            println!("hello\t:print Hello, World!");
-            println!("reboot\t:reboot the device");
-            println!("ls\t:list files in the initramfs");
-            println!("cat\t:print the content of a file in the initramfs");
-            println!("exec\t:load a file to memory and execute it");
+            commands::help();
         } else if inp_buf.starts_with(b"hello") {
             println!("Hello, World!");
         } else if inp_buf.starts_with(b"reboot") {
@@ -45,19 +71,9 @@ pub fn start(initrd_start: u32) {
                 }
             }
 
-            unsafe {
-                asm!(
-                    "mov {tmp}, 1",
-                    "msr cntp_ctl_el0, {tmp}", // Enable the timer
-                    "mrs {tmp}, cntfrq_el0",
-                    "msr cntp_tval_el0, {tmp}",
-                    "mov {tmp}, 2",
-                    "ldr {tmp2}, =0x40000040", // CORE0_TIMER_IRQ_CTRL
-                    "str {tmp}, [{tmp2}]",
-                    tmp = out(reg) _,
-                    tmp2 = out(reg) _,
-                );
+            timer::add_timer_ms(2000, Box::new(|| set_next_timer()));
 
+            unsafe {
                 asm!(
                     "mov {tmp}, 0x200",
                     "msr spsr_el1, {tmp}",
@@ -68,13 +84,34 @@ pub fn start(initrd_start: u32) {
                     "eret",
                     tmp = out(reg) _,
                 );
+
+                println!("Should not reach here");
             }
         } else if inp_buf.starts_with(b"M3") {
             println!("香的發糕");
+
+        } else if inp_buf.starts_with(b"setTimeout") {
+            let inp_buf = &inp_buf[..len - 1];
+            let mut iter = inp_buf.split(|c| c.clone() == b' ');
+            assert_eq!(iter.next().unwrap(), b"setTimeout");
+            let message = core::str::from_utf8(iter.next().unwrap()).unwrap().to_string();
+            let time = core::str::from_utf8(iter.next().unwrap()).unwrap();
+            match time.parse::<u64>() {
+                Ok(time) => {
+                    timer::add_timer_ms(time, move || println!("{}", message));
+                }
+                Err(_) => {
+                    println!("Invalid time: {}", time.len());
+                }
+            }
+            
+            // timer::add_timer_ms(time, move || println!("{}", message));
+        
+        } else if inp_buf.starts_with(b"test_memory") {
+            commands::test_memory();
         } else {
             println!(
-                "Unknown command {}",
-                core::str::from_utf8(&inp_buf).unwrap().trim_end()
+                "Unknown command"
             );
         }
     }
